fix typo & rename variables
