+++ src/components/CommentPanel/ColorPicker.vue
